Add files via upload
IteminBasket is the main program.  Using this file we can select items and the total cost is calculated.

Further, It uploads the data from the CSV file using the class file 'uploading.java'.  'Basket.java' is a class file and keeps the items that are added to the basket.

Future:  Basket delete option.  Updating stock in csv file.  Creating a log file and an save old version of CSV file.